refactor: moved state management to custom hook useOpeningState
